+++ src/MapPoint.hpp
@@ -22,22 +22,26 @@ public: // public members
     MapPoint();
     /**
      * @brief Construct map point with input coordinate and descriptor.
-     * @param[in] X      3D world coordinate as a \f$3 \times 1\f$ matrix.
-     * @param[in] desc   Descriptor of the map point.
-     * @param[in] frmIdx Index of the frame that observed the point.
+     * @param[in] X3D     3D world coordinate as a \f$3 \times 1\f$ matrix.
+     * @param[in] desc    Descriptor of the map point.
+     * @param[in] nIdxFrm Index of the frame that observed the point.
      */
-    MapPoint(const cv::Mat& X, const cv::Mat& desc, int frmIdx);
+    MapPoint(const cv::Mat& X3D, const cv::Mat& desc, int nIdxFrm);
     /**
      * @name Class Getter/Setters
      * @brief Basic getters/setters for read/update various info on the point.
      */
     ///@{
     /// Get 3D world coordinate.
-    cv::Mat getPt() const { return mX; }
+    cv::Mat getX3D() const { return mX3D; }
     /// Get descriptor of the point.
     cv::Mat getDesc() const { return mDesc; }
+    /// Get index of the latest frame that observed the point.
+    int getIdxLastObsFrm() const { return mnIdxLastObsFrm; }
+    /// Get match-to-observation ratio.
+    float getMatch2ObsRatio() const;
     /// Update 3D world coordinate of the point.
-    void setPt(const cv::Mat& X) { mX = X.clone(); }
+    void setX3D(const cv::Mat& X3D) { mX3D = X3D.clone(); }
     /// Update descriptor of the point.
     void setDesc(const cv::Mat& desc) { mDesc = desc.clone(); }
     /// Update index of the latest frame that observed the point.
@@ -49,7 +53,7 @@ public: // public members
     void addCntMatches(int n = 1);
 private: // private data
     /// Inhomogeneous 3D world coordinate of the point.
-    cv::Mat mX;
+    cv::Mat mX3D;
     /// Best descriptor of the point.
     cv::Mat mDesc;
     /// Index of the latest frame by which the point is observed.